higherorderplach

+++ extended_distributed_two.py
@@ -424,8 +424,8 @@ class All2All_Req(Function):
                     a2a_info.global_table_num
                     * a2a_info.local_batch_num
                     * a2a_info.emb_dim
-                ]
-            )
+                ], dtype = torch.float64 
+            ) 
             req = dist.all_to_all_single(
                 output, input, table_split_lengths, batch_split_lengths, async_op=True
             )
